+++ demos/employees/src/main/java/is/codion/demos/employees/ui/EmployeeEditPanel.java
@@ -72,7 +72,7 @@ public class EmployeeEditPanel extends EntityEditPanel {
 		EntityComboBox departmentBox = (EntityComboBox) component(Employee.DEPARTMENT_FK).get();
 		NumberField<Integer> departmentNumberField = departmentBox.integerSelectorField(Department.DEPARTMENT_NO)
 						.transferFocusOnEnter(true)
-						.validIndicator(editModel().editor().value(Employee.DEPARTMENT_FK).valid())
+						.valid(editModel().editor().value(Employee.DEPARTMENT_FK).valid())
 						.build();
 
 		component(Employee.DEPARTMENT_FK).replace(departmentNumberField);