+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/TemporalFieldPanel.java
@@ -23,7 +23,7 @@ import is.codion.swing.common.ui.component.Components;
 import is.codion.swing.common.ui.component.builder.AbstractComponentValueBuilder;
 import is.codion.swing.common.ui.component.builder.ComponentValueBuilder;
 import is.codion.swing.common.ui.component.calendar.CalendarPanel;
-import is.codion.swing.common.ui.component.indicator.ValidIndicatorFactory;
+import is.codion.swing.common.ui.component.indicator.ValidIndicator;
 import is.codion.swing.common.ui.component.value.AbstractComponentValue;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 import is.codion.swing.common.ui.key.TransferFocusOnEnter;
@@ -296,9 +296,9 @@ public final class TemporalFieldPanel<T extends Temporal> extends JPanel {
 		}
 
 		@Override
-		protected void enableValidIndicator(ValidIndicatorFactory validIndicatorFactory,
+		protected void enableValidIndicator(ValidIndicator validIndicator,
 																				TemporalFieldPanel<T> component, ObservableState valid) {
-			validIndicatorFactory.enable(component.temporalField, valid);
+			validIndicator.enable(component.temporalField, valid);
 		}
 
 		private TemporalField<T> createTemporalField() {