+++ swing/common-ui/src/main/java/module-info.java
@@ -81,7 +81,7 @@ module is.codion.swing.common.ui {
 
 	uses is.codion.swing.common.ui.laf.LookAndFeelProvider;
 	uses is.codion.swing.common.ui.component.indicator.ModifiedIndicatorFactory;
-	uses is.codion.swing.common.ui.component.indicator.ValidIndicatorFactory;
+	uses is.codion.swing.common.ui.component.indicator.ValidIndicator;
 	uses is.codion.swing.common.ui.scaler.Scaler;
 
 	provides is.codion.swing.common.ui.laf.LookAndFeelProvider
@@ -89,8 +89,8 @@ module is.codion.swing.common.ui {
 	provides is.codion.swing.common.ui.component.indicator.ModifiedIndicatorFactory
 					with is.codion.swing.common.ui.component.indicator.UnderlineModifiedIndicatorFactory,
 									is.codion.swing.common.ui.component.indicator.SwapColorsModifiedIndicatorFactory;
-	provides is.codion.swing.common.ui.component.indicator.ValidIndicatorFactory
-					with is.codion.swing.common.ui.component.indicator.BackgroundColorValidIndicatorFactory;
+	provides is.codion.swing.common.ui.component.indicator.ValidIndicator
+					with is.codion.swing.common.ui.component.indicator.BackgroundColorValidIndicator;
 	provides is.codion.swing.common.ui.scaler.Scaler
 					with is.codion.swing.common.ui.scaler.FontSizeScaler;
 }